remove service interfaces initialize list in recommendation request entity change endpoints path
