+++ src/main/java/school/faang/user_service/mapper/RecommendationRequestMapper.java
@@ -10,7 +10,6 @@ import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
 import school.faang.user_service.entity.recommendation.SkillRequest;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Mapper(componentModel = "spring")
@@ -33,9 +32,6 @@ public interface RecommendationRequestMapper {
     GetRecommendationRequestResponse toGetDto(RecommendationRequest recommendationRequest);
 
     default List<String> toSkills(List<SkillRequest> skillRequests) {
-        if (skillRequests == null) {
-            return new ArrayList<>();
-        }
         return skillRequests.stream()
                 .map(SkillRequest::getSkill)
                 .map(Skill::getTitle)
