--- src/main/java/school/faang/user_service/service/recommendation/SkillService.java
@@ -1,8 +1,0 @@
-package school.faang.user_service.service.recommendation;
-
-import school.faang.user_service.entity.Skill;
-
-public interface SkillService {
-
-    Skill findById(long id);
-}
