--- src/main/java/school/faang/user_service/service/recommendation/UserService.java
@@ -1,8 +1,0 @@
-package school.faang.user_service.service.recommendation;
-
-import school.faang.user_service.entity.User;
-
-public interface UserService {
-
-    User findById(long id);
-}
